requirements.txt
absl-py==2.1.0
addict==2.4.0
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
codeformer==0.0.11
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dlib @ file:///C:/Users/Admin/Downloads/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
easydict==1.13
face-recognition==1.3.0
face_recognition_models==0.3.0
facelib @ git+https://github.com/sajjjadayobi/FaceLib.git@382841efec823f059f49754e5d378abb12cfb551
facexlib==0.3.0
filelock==3.16.0
filterpy==1.4.5
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.9.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.66.1
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.0
inflect==7.3.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pilgram==1.2.1
pillow==10.4.0
platformdirs==4.3.2
protobuf==5.28.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sympy==1.13.2
tb-nightly==2.18.0a20240913
tbb==2021.13.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
tomli==2.0.1
torch==2.4.1+cpu
torchaudio==2.4.1+cu118
torchvision==0.19.1+cpu
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.44.0
yapf==0.40.2
zipp==3.20.1
